add support for sh file

+++ main.go
@@ -18,10 +18,13 @@ import (
 // this var used to store the shortcut file path, taken from argument during `go run`
 var shortcutFilePath = ""
 
-// these vars below will have values injected during build time
+// these vars below will have values injected during build time, for .exe creation
 var isRuntime = ""
 var url = ""
 var icon = ""
+
+// this var below will be used to store the bash script command, for .sh creation
+var bashScriptCommand = ""
 
 func main() {
 
@@ -49,53 +52,68 @@ func buildExecutable() {
 	// construct the executable file namen
 	filename := filepath.Base(shortcutFilePath)
 	fileExt := strings.ToLower(filepath.Ext(filename))
-	filename = filename[0:len(filename)-len(fileExt)] + ".exe"
-
-	if fileExt != ".lnk" && fileExt != ".url" && fileExt != ".cda" && fileExt != ".desktop" {
+	shouldBecomeDotExe := true
+	if fileExt == ".lnk" || fileExt == ".url" || fileExt == ".cda" {
+		filename = filename[0:len(filename)-len(fileExt)] + ".exe"
+	} else if fileExt == ".desktop" {
+		filename = filename[0:len(filename)-len(fileExt)] + ".sh"
+		shouldBecomeDotExe = false
+	} else {
 		forceExitIfError("unsupported file. file must be one of these: .lnk, .url, .cda, .desktop")
 	}
 
 	// get file metadata
 	loadShortcutFileMetadata()
 
-	// fail if gopath is not set
-	if os.Getenv("GOPATH") == "" {
-		forceExitIfError("GOPATH env var is required to be set")
-	}
-
-	// fail if go binary is not callable
-	if _, err := exec.LookPath("go"); err != nil {
-		forceExitIfError("the go binary is required to be added into PATH env var")
-	}
-
-	// fail if git binary is not calleble, required by `go get` command
-	if _, err := exec.LookPath("git"); err != nil {
-		forceExitIfError("the git binary is required to be added into PATH env var")
-	}
-
-	// go get the rsrc lib
-	execCommand([]string{"go", "get", "github.com/akavel/rsrc"})
-
-	// create the rsrc.syso to set the icon of upcoming executable
-	gopath := os.Getenv("GOPATH")
-	rsrcFilename := "rsrc.syso"
-	rsrcExecFile := filepath.Join(gopath, "bin", strings.Split(rsrcFilename, ".")[0])
-	execCommand([]string{rsrcExecFile, "-ico", icon})
-
-	// remove existing file
-	os.Remove(filename)
-
 	// build the executable
-	execCommand([]string{
-		"go",
-		"build",
-		"-ldflags", fmt.Sprintf(`-X "main.isRuntime=true" -X "main.url=%s" -X "main.icon=%s"`, url, icon),
-		"-o", filename,
-	})
-	fmt.Printf("  => result: executable %s is successfully generated", filename)
+	//   => .lnk, .url, .cda will be converted into .exe
+	//   => .desktop will be converted into .sh
+	if shouldBecomeDotExe {
 
-	// remove syso file
-	os.Remove(rsrcFilename)
+		// fail if gopath is not set
+		if os.Getenv("GOPATH") == "" {
+			forceExitIfError("GOPATH env var is required to be set")
+		}
+
+		// fail if go binary is not callable
+		if _, err := exec.LookPath("go"); err != nil {
+			forceExitIfError("the go binary is required to be added into PATH env var")
+		}
+
+		// fail if git binary is not calleble, required by `go get` command
+		if _, err := exec.LookPath("git"); err != nil {
+			forceExitIfError("the git binary is required to be added into PATH env var")
+		}
+
+		// go get the rsrc lib
+		execCommand([]string{"go", "get", "github.com/akavel/rsrc"})
+
+		// create the rsrc.syso to set the icon of upcoming executable
+		gopath := os.Getenv("GOPATH")
+		rsrcFilename := "rsrc.syso"
+		rsrcExecFile := filepath.Join(gopath, "bin", strings.Split(rsrcFilename, ".")[0])
+		execCommand([]string{rsrcExecFile, "-ico", icon})
+
+		// remove existing file
+		os.Remove(filename)
+
+		// build the .exe file
+		execCommand([]string{
+			"go",
+			"build",
+			"-ldflags", fmt.Sprintf(`-X "main.isRuntime=true" -X "main.url=%s" -X "main.icon=%s"`, url, icon),
+			"-o", filename,
+		})
+
+		// remove syso file
+		os.Remove(rsrcFilename)
+	} else {
+		content := []byte(fmt.Sprintf("#!/bin/sh\n\n%s", bashScriptCommand))
+		err := ioutil.WriteFile(filename, content, os.ModePerm)
+		forceExitIfError(err)
+	}
+
+	fmt.Printf("  => result: executable %s is successfully generated", filename)
 }
 
 // get metadata of shortcut file
@@ -121,6 +139,7 @@ func loadShortcutFileMetadata() {
 
 	url = lines["url"]
 	icon = lines["icon"]
+	bashScriptCommand = lines["exec"]
 }
 
 // utility func to force fail whenever there is error occuring
